requirements.txt
chainlit
python-dotenv
fastapi
uvicorn
google-generativeai
chromadb
docker
pandas
matplotlib
seaborn 
pillow